fstab, initscripts: Fixup the /var layout
Modern Linux systems have a volatile /run directory and /var/run links to that. Add
a mountpoint in fstab to accomodate that, copied from the standard base-files.

There's a cryptic "devusbfs" entry in fstab that doesn't appear to have any effect
since most kernels run without /proc/bus/usb existing anyway, so remove the line.

Remove the 'volatiles' entry in initscripts and use the one from OE-core, so that
we get the standard layout of /var and its various links to /run.

The only real change to /var was to have /var/cache in volatile as well, arrange that
in the initscript bbappend instead of replacing its volatiles list.